+++ src/bemserver_api/resources/st_cleanups_by_timeseries/__init__.py
@@ -1,5 +1,0 @@
-from .routes import blp  # noqa
-
-
-def register_blueprints(api):
-    api.register_blueprint(blp)